agentcore_deployment/orchestrator_agent/requirements.txt
# Minimal requirements for AgentCore and MCP
bedrock-agentcore
strands-agents
boto3>=1.40.0
pydantic>=2.11.0

# MCP Support
mcp>=1.10.0
fastapi>=0.104.0
starlette>=0.27.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Observability
aws-opentelemetry-distro>=0.10.1